Update to a special role resource.
Create a new role resource that ties into the Vault API to set AWS IAM
policies. This could be done using vault_generic_secret before, but
making it its own resource makes it more user-friendly and intuitive,
and allows us to infer the path.

Update the data source to accept a backend path, a role name, and a type
of credentials instead of forcing users to hardcode paths.